blacklistfilter: moved sources from subdirs
