--- blacklistfilter/urldetect/bld_userConfigFile.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <struct name="main struct">
-        <!-- Array with information about public blacklist -->
-        <array name="blacklist_arr">
-            <struct>
-                <!-- ID of the blacklist, blacklisted flows are flagged with corresponding ID of blacklist
-                     BEWARE: Could be number from interval <0, 63> -->
-                <element name="id">0</element>
-                <!-- Method of retrieving blacklist, currently supported is only web via wget tool -->
-                <element name="method">web</element>
-                <!-- Name of the blacklist, modules uses this name to choose which blacklist to use -->
-                <element name="name">Malware Domains</element>
-                <!-- Address from which the blacklist will be downloaded -->
-                <element name="source">http://mirror2.malwaredomains.com/files/justdomains</element>
-                <!-- Type of blacklist, only informational value -->
-                <element name="type">BOTNET</element>
-                <!-- Regular expression to extract needed information (IP for example) from blacklist source -->
-                <element name="regex">.*</element>
-            </struct>
-            <struct>
-                <element name="id">6</element>
-                <element name="method">web</element>
-                <element name="source">http://data.phishtank.com/data/online-valid.csv</element>
-                <element name="name">PhishTank</element>
-                <element name="type">Phishing</element>
-                <element name="regex">.*</element>
-            </struct>
-            <struct>
-                <element name="id">9</element>
-                <element name="method">web</element>
-                <element name="source">http://84.200.52.65/dev.booterblacklist/data/booterlist_latest.txt</element>
-                <element name="name">Booter</element>
-                <element name="type">DDOS</element>
-                <element name="regex">.*</element>
-            </struct>
-        </array>
-    </struct>
-</configuration>
